Better source tree and fixed some warnings.
Fixed warning at line 130 of the file `sha3.c` which was caused by a
comparison between an `unsigned int` counter on the for loop and the
`int` that specified the size of hash.

The solution was to use `strtoul` at line 64 which is guaranteed to read 64bits
and then cast it back to an unsigned integer.

This can be further improved by using `unsigned short`.

---

Fixed a warning on the `sha3.h` file since the header `stddef.h`
required by `size_t` was not included.

---

Refactored the Makefile to match the new source tree.